AttentiveRanker/src/requirements.txt
absl-py==0.8.0
astor==0.8.0
gast==0.3.1
grpcio==1.23.0
h5py==2.10.0
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
mock==3.0.5
numpy==1.15.4
pkg-resources==0.0.0
protobuf==3.9.1
PyYAML==5.1.2
scipy==1.3.1
six==1.12.0
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
termcolor==1.1.0
tqdm==4.35.0
Werkzeug==0.15.6
